Compile all templates using $compile as part of the unit test.
This ensures that, at least for flags used in jade templates, all
flags are specified as part of the unit test step at release. Holds
off the need for e2e tests for right now.

N.B. prx-flag directives in index.html will not be checked until
an e2e suite is in place.