Support linking to models via `ui-sref="model"`
Rather than encoding the state names and parameters in different modules,
this allows a default implementation (state name is `<modelType>.show`,
state params are `{<modelType>Id: <model>.id}`) and the ability to
override at the model level. Because models are typically defined in the
same module as their states, this reflects the structure of the app better
and allows more localized management of a given module.

```html
<a ng-href="/accounts/{{account.id}}"></a>
<a ui-sref="account.show({accountId:account.id})"></a>
<a ui-sref="account"></a>
```

These are generally equivalent, but the first becomes out of date if
the state's url format changes, the second becomes out of date if the
state's name (might happen if heirarchy needs to change) or parameters
change.