Use `Controller as x` syntax
By using controller as x syntax, it becomes easier to reason about
which part of the app is responsible for which pieces of scope when
nesting controllers (as we begin doing soon for things like stories
within accounts). It also removes the temptation to communicate
between controllers using $scope, which always leads to heartache.