Use semantic markup, break module concerns out
* Use the semantic markup patterns agreed upon in chat room
* Break module-specific styles into module specific stylesheets
* Use soft-dependencies on external modules for rendering (see below)
* remove crufty styles we are no longer using

If styles are split by the module using them, it becomes important
to ensure that the stories module is responsible for rendering stories.
Using directives can accomplish this pretty easily, and declaratively.

By declaring a directive with an isolate scope using a template, we're
effectively creating a clean environment in which the module can render
however it needs to. The accounts module doesn't need to be aware of what
methods the stories module adds to story objects, because it (very likely)
never needs to call them.