Add low res versions of images that were missing
This should be automated in the future. The way that I
imagine it working is:

  1. Grab all images which have the @2x suffix
  2. Filter out all images which have a non-suffixed version in the src
     directory that is newer than the @2x version.
  3. Print a warning for all non-suffixed images that are older than the @2x versions.
  4. Generate a half-sized image with the suffix removed and put it
     in the destination directories. Do not modify anything in the src
     directory.

The benefits are that we still have a way to generate a low-res version manually
but we don't have to when it's not completely necessary.

This will have to be part of the gulp restructuring alluded to elsewhere.