Move installation and testing concerns into NPM
Makes npm responsible for installing bower packages and running
gulp. The primary benefit is that global packages are no longer
needed since the local npm bundle is automatically placed in path
in this context.

Additionally, by moving to a more standard process for installing
and starting tests, we're simplifying the responsibilities of the
fabled E2E CI/CD system.