Normalize negative listen durations, as well.
There's a (small) chance that some of the extremely high values are so
high because Google Analytics converts from a signed integer to an
unsigned integer, making a negative number into an extremely large positive
number.

This seeks to address that. I added a small window (of under 0.5 seconds)
because it felt right to do so. Not sure if it actually makes sense.