Render picks nicely on wide views of the home page.
The most interesting bit is definitely the system for monitoring if elements
are overflowing or not and calling back. I can walk through how it works
because it's a little obtuse, but the reason is that we want to avoid DOM
thrashing (reading and writing to the DOM) so we batch our reads and then loop
through to see if anything changed and call those things back. It's a little
like the angular digest cycle in that way.