Make nav structure registrable throughout the app
This is still a WIP so there are likely some issues or changes needed to the
API, but the basic idea is that the module which relates to the specific piece
of the nav/drawer system registers it along with preferences about how it is
ordered and whether or not it should be preferentially kept in the nav as space
is constrained.

At present we have both types of sorting in place. This is largely influenced by
the Android system for registering the Menu in the ActionBar, but much less
capable (currently only supports the things which are already reflected by the
designs).

Next steps are to get what android calls "ActionProvider"s in place which in
this context will just mean directives which can be swapped in either on
activation (as with the search item) or at all times (as with the login widget).