Automatically flex arrangement of nav/drawer items
Automatically move things between the drawer and the top nav as more or less
space is available in the nav bar.

As we change the global nav, this allows us to set priorities and sorting order
of those nav elements and have it scale to the various screen sizes
automatically. The use case currently has only four breakpoints, which are:

  * empty nav (search, browse, login in drawer)
  * nav with login only (search, browse in drawer)
  * nav with search and login (browse in drawer)
  * nav with search, browse, and login (empty drawer)

We have enough room for more things to go into the nav if we want to, so that
would mean more break points. Additionally, the sorting and nav priority is
handled automatically.